Moving files and some cleanup
