+++ Anima/segmentation/CMakeLists.txt
@@ -7,3 +7,4 @@ project(ANIMA-SEGMENTATION)
 add_subdirectory(basic_tools)
 add_subdirectory(graph_cut)
 add_subdirectory(isosurface)
+add_subdirectory(segmentation_performance_analyzer)
