+++ Anima/segmentation/basic_tools/CMakeLists.txt
@@ -8,5 +8,4 @@ add_subdirectory(mask_image)
 add_subdirectory(morphological_operations)
 add_subdirectory(otsu_thr_image)
 add_subdirectory(thr_image)
-add_subdirectory(segmentation_performance)
 add_subdirectory(total_lesion_load)
